fix: 拦截后置 Todo 创建声明
